Fixed bug with read-until's implementation of the KMP algorithm.
The KMP algorithm works by building a table which keeps track of which
characters in the search string can be skipped when a mismatch
occurs. This information is supposed to be maintained in the index of
the mismatch. Read-until had previously built the table incorrectly
and placed that information in the wrong place.