Fixed bug with how read-until uses mismatch.
When one of the strings passed to mismatch is a proper prefix of the
other, mismatch will return the last index tested in the first string
plus one. This means that when the first string is a proper prefix of
the second string mismatch will return the length of the first
string. To avoid this case, just test that the index of the mismatch
is less than the length of the string.