Escape wildcard characters in path names.
Changed parse-path to use make-pathname for automatic implementation-specific escaping of wildcard characters in path names.

Fixes an issue on SBCL where an unmatched '[' character in the URL causes an Internal Server Error, and if it's matched with a ']' it causes a Not Found. Because SBCL treats un-escaped square brackets in pathnames as wildcards.