Fix shutdown acceptor wake when listening on 0.0.0.0.
0.0.0.0/8 (and the IPv6 equivalent, ::) are invalid addresses to use as
the target of a connection, but those are the local addresses of an
acceptor listening on them. In order for WAKE-ACCEPTOR-FOR-SHUTDOWN to
connect to such an acceptor, we need to use an alternate address when
that's the case.

This also adds a test to detect the failure to interrupt
USOCKET:WAIT-FOR-INPUT which causes the hang. This runs in the same test
context as the normal confidence tests (to preserve the summary at the
end of the run), but uses its own acceptor to avoid potential state
issues. Test-failure conditions have been refactored slightly to support
more general tests like this.