Avoid setting *DEFAULT-PATHNAME-DEFAULTS* to a file pathname during tests.
When *D-P-D* is a file pathname, pathnames for files that do not have a
file extension wind up having an extension added by MERGE-PATHNAMES, even
when that is only intended to merge the path component (this may depend on
whether the implementation uses NIL or :UNSPECIFIC as the :TYPE component
when parsing paths that don't have an extension).

In particular, this causes a problem when drakma's user-agent-string code
calls SOFTWARE-VERSION on linux; SBCL's implementation reads the
/proc/sys/kernel/osrelease file, but MERGE-PATHNAMES causes it to try to
open /proc/sys/kernel/osrelease.lisp, which doesn't exist. The test need
the directory component of *D-P-D* set for some file-handling tests, but
they do not rely on the filename portion, so strip that before rebinding
*D-P-D* around the tests.