Ensure that an EASY-ACCEPTOR is composable with other acceptors
The ACCEPTOR-DISPATCH-REQUEST returns the values resulting from
a call to CALL-NEXT-METHOD, to ensure EASY-ACCEPTOR is composable
with other acceptors, implementing other ACCEPTOR-DISPATCH-REQUEST
methods.  The value returned is important to PROCESS-REQUEST which
uses it to determine if the handler used by ACCEPTOR-DISPATCH-REQUEST
prepared an in-memory version of the answer or wrote it in
the appropriate output stream.